changed to double

+++ src/search.cpp
@@ -1837,7 +1837,7 @@ void update_continuation_histories(Stack* ss, Piece pc, Square to, int bonus) {
     constexpr std::array<ConthistBonus, 6> conthist_bonuses = {
       {{1, 1025}, {2, 621}, {3, 325}, {4, 512}, {5, 122}, {6, 534}}};
 
-    int multiplier = bonus / 1024;
+    double multiplier = bonus / 1024;
 
     for (const auto [i, weight] : conthist_bonuses)
     {
